[libclang] Split-out parts of `Index.h`
The `Index.h` header defines some functions and data structures that are useful for other (non-indexing) APIs. This patch extracts those into separate headers.

Reviewed By: akyrtzi, benlangmuir

Differential Revision: https://reviews.llvm.org/D134837

(cherry picked from commit 29e0435ac04957861aa1f85d41291c8b19db0122)
